docs: implement a way to automatically generate API docs
