+++ api/docs/README.md
@@ -1,0 +1,11 @@
+count-up-down
+
+# count-up-down
+
+## Table of contents
+
+### Modules
+
+- [common/calculate](modules/common_calculate.md)
+- [common/types](modules/common_types.md)
+- [node](modules/node.md)
