+++ api/docs/interfaces/common_types.CountResult.md
@@ -1,0 +1,79 @@
+[count-up-down](../README.md) / [common/types](../modules/common_types.md) / CountResult
+
+# Interface: CountResult
+
+[common/types](../modules/common_types.md).CountResult
+
+This is the data structure of the raw numbers that we want to show on the screen.
+The numbers can be negative (count down) or positive (count up).
+
+## Table of contents
+
+### Properties
+
+- [days](common_types.CountResult.md#days)
+- [hours](common_types.CountResult.md#hours)
+- [minutes](common_types.CountResult.md#minutes)
+- [months](common_types.CountResult.md#months)
+- [seconds](common_types.CountResult.md#seconds)
+- [years](common_types.CountResult.md#years)
+
+## Properties
+
+### days
+
+• **days**: `number`
+
+#### Defined in
+
+[common/types.ts:8](https://github.com/imballinst/count-up-down/blob/4ec7333/src/common/types.ts#L8)
+
+___
+
+### hours
+
+• **hours**: `number`
+
+#### Defined in
+
+[common/types.ts:9](https://github.com/imballinst/count-up-down/blob/4ec7333/src/common/types.ts#L9)
+
+___
+
+### minutes
+
+• **minutes**: `number`
+
+#### Defined in
+
+[common/types.ts:10](https://github.com/imballinst/count-up-down/blob/4ec7333/src/common/types.ts#L10)
+
+___
+
+### months
+
+• **months**: `number`
+
+#### Defined in
+
+[common/types.ts:7](https://github.com/imballinst/count-up-down/blob/4ec7333/src/common/types.ts#L7)
+
+___
+
+### seconds
+
+• **seconds**: `number`
+
+#### Defined in
+
+[common/types.ts:11](https://github.com/imballinst/count-up-down/blob/4ec7333/src/common/types.ts#L11)
+
+___
+
+### years
+
+• **years**: `number`
+
+#### Defined in
+
+[common/types.ts:6](https://github.com/imballinst/count-up-down/blob/4ec7333/src/common/types.ts#L6)
