+++ api/docs/modules/common_types.md
@@ -1,0 +1,9 @@
+[count-up-down](../README.md) / common/types
+
+# Module: common/types
+
+## Table of contents
+
+### Interfaces
+
+- [CountResult](../interfaces/common_types.CountResult.md)
