+++ api/docs/modules/node.md
@@ -1,0 +1,15 @@
+[count-up-down](../README.md) / node
+
+# Module: node
+
+## Table of contents
+
+### References
+
+- [calculate](node.md#calculate)
+
+## References
+
+### calculate
+
+Re-exports [calculate](common_calculate.md#calculate)
