+++ package.json
@@ -3,15 +3,15 @@
   "version": "0.1.2",
   "jsdelivr": "dist/count-up-down.min.js",
   "browser": "dist/count-up-down.min.js",
-  "main": "dist/cjs/index-node.js",
-  "esnext": "dist/esm/index-node.js",
-  "module": "dist/esm/index-node.js",
-  "types": "dist/types/index-node.d.ts",
-  "typings": "dist/types/index-node.d.ts",
+  "main": "dist/cjs/node/index.js",
+  "esnext": "dist/esm/node/index.js",
+  "module": "dist/esm/node/index.js",
+  "types": "dist/types/node/index.d.ts",
+  "typings": "dist/types/node/index.d.ts",
   "exports": {
     ".": {
-      "require": "./dist/cjs/index-node.js",
-      "default": "./dist/esm/index-node.js"
+      "require": "./dist/cjs/node/index.js",
+      "default": "./dist/esm/node/index.js"
     }
   },
   "repository": "git@github.com:imballinst/count-up-down.git",
@@ -24,7 +24,8 @@
     "build:types": "tsc --project tsconfig.build-esm.json --emitDeclarationOnly --declaration --declarationDir dist/types",
     "bundle": "ts-node scripts/build/index.ts",
     "test": "jest",
-    "test:watch": "jest --watch"
+    "test:watch": "jest --watch",
+    "generate:doc": "rm -rf api/docs && typedoc --plugin typedoc-plugin-markdown --readme none --out api/docs src/node src/browser src/common/*.ts"
   },
   "devDependencies": {
     "@babel/core": "7.15.5",
@@ -38,6 +39,8 @@
     "esbuild": "0.12.24",
     "jest": "27.1.0",
     "ts-node": "10.2.1",
+    "typedoc": "0.22.7",
+    "typedoc-plugin-markdown": "3.11.3",
     "typescript": "4.4.2"
   },
   "browserslist": [
