+++ src/common/types.ts
@@ -1,0 +1,12 @@
+/**
+ * This is the data structure of the raw numbers that we want to show on the screen.
+ * The numbers can be negative (count down) or positive (count up).
+ */
+export interface CountResult {
+  years: number;
+  months: number;
+  days: number;
+  hours: number;
+  minutes: number;
+  seconds: number;
+}
