--- src/index-node.ts
@@ -1,1 +1,0 @@
-export { calculate } from './calculate';
