+++ src/node/index.ts
@@ -1,0 +1,1 @@
+export { calculate } from '../common/calculate';
