+++ tsconfig.json
@@ -12,7 +12,7 @@
     "resolveJsonModule": true,
     "isolatedModules": true
   },
-  "include": ["./src/*.ts", "./scripts/**/*.ts"],
+  "include": ["src", "scripts"],
   "exclude": ["node_modules"],
   "ts-node": {
     "compilerOptions": {
